Move cluster class to own module
